app/Executable/parquet2mysql/requirements.txt
polars
pyarrow
PyMySQL
SQLAlchemy
sqlalchemy-utils
PyYAML